basicwidget: change LineHeight to integer
A float line height caused an unexpected height mismatch between
Select and Button widgets, as the Measure results depend on how the
value is rounded. To avoid confusion, use integers instead.

+++ basicwidget/form.go
@@ -112,7 +112,7 @@ func (f *Form) appendItemBounds(context *guigui.Context, itemBounds []image.Rect
 		}
 		itemBounds = append(itemBounds, b)
 
-		maxPaddingY := paddingS.Y + int((float64(u)-LineHeight(context))/2)
+		maxPaddingY := paddingS.Y + (u-LineHeight(context))/2
 		if item.PrimaryWidget != nil {
 			bounds := b
 			bounds.Min.X += paddingS.X
